+++ App/backend/api/admin.py
@@ -1,5 +1,6 @@
 from django.contrib import admin
-from .models import User, Tag, Comment, ArtItem
+from .models.models import Tag, Comment, ArtItem
+from .models.user import User
 
 
 class UserAdmin(admin.ModelAdmin):
