+++ App/backend/api/serializers/serializers.py
@@ -7,7 +7,9 @@ from .models import User, Tag, Comment, ArtItem
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
         model = User
-        fields = ['id', 'username',  'is_active_user', 'name', 'surname', 'email', 'profile_image', 'created_at', 'updated_at']
+        fields = ['id', 'username',  'is_active_user', 'name', 'surname',
+                  'email', 'profile_image', 'created_at', 'updated_at']
+
 
 class TagSerializer(serializers.ModelSerializer):
     class Meta:
@@ -18,14 +20,18 @@ class TagSerializer(serializers.ModelSerializer):
 class ArtItemSerializer(serializers.ModelSerializer):
     class Meta:
         model = ArtItem
-        fields = ['id', 'title', 'description', 'owner', 'tags', 'artitem_image']
-    
+        fields = ['id', 'title', 'description',
+                  'owner', 'tags', 'artitem_image']
+
     def to_representation(self, instance):
         rep = super().to_representation(instance)
         rep["tags"] = TagSerializer(instance.tags.all(), many=True).data
         return rep
 
+
 class CommentSerializer(serializers.ModelSerializer):
     class Meta:
         model = Comment
         fields = ['id', 'body', 'commented_by', 'commented_on', 'created_at']
+
+
